+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Unreleased
 
 * Add an option for title language to public layout ([PR #5169](https://github.com/alphagov/govuk_publishing_components/pull/5169))
+* Remove govspeak chart's hidden text and update button text ([PR #5165](https://github.com/alphagov/govuk_publishing_components/pull/5165))
 
 ## 62.2.1